fixed linux

+++ MenuMode.cpp
@@ -138,8 +138,8 @@ void MenuMode::draw(glm::uvec2 const& drawable_size) {
         glBindVertexArray(VAO);
 
         if (c.line!=line){
-            cursor_x = drawable_size.x/2.0*c.start_x;
-            cursor_y = drawable_size.y/2.0*c.start_y+c.line*y_offset;
+            cursor_x = (float)drawable_size.x/2.0f*c.start_x;
+            cursor_y = (float)drawable_size.y/2.0f*c.start_y+c.line*y_offset;
             line = c.line;
         }
